Modificaciones en cita/models.py
