+++ registro/modelos/descripcion_models.py
@@ -44,9 +44,41 @@ class Terapia(models.Model):
                        (7, "TERAPIA FAMILIAR"),
                        (8, "NINGUNA"))
     tipo = models.SmallIntegerField("Tipo de terapia", choices=TERAPIA_CHOICES)
+    costo_sesion = models.PositiveSmallIntegerField(default=1, blank=False, null=False)    
     tiempo_terapia = models.DurationField("¿Cuánto tiempo lleva realizando la terapia")
     descripcion = models.ForeignKey(Descripcion)
-    
+
+
+class Terapista(models.Model):
+    """ Representa el terapista del paciente """
+    GRUPO_SANGUINEO_CHOICES = (
+        ("A+", "A+"),
+        ("A-", "A-"),
+        ("O+", "O+"),
+        ("O-", "O-"),
+        ("B+", "B+"),
+        ("B-", "B-"),
+        ("AB+", "AB+"), 
+        ("AB-", "AB-")        
+    )
+    cedula = models.CharField(max_length=10)
+    nombres = models.CharField(max_length=50)
+    apellidos = models.CharField(max_length=50)
+    direccion = models.CharField(max_length=256)
+    telefonos = models.CharField(max_length=50)
+    fecha_nacimiento = models.DateField("fecha de nacimiento")
+    grupo_sanguineo = models.CharField("grupo sanguineo",
+                                        choices=GRUPO_SANGUINEO_CHOICES,
+                                        max_length=4)
+    terapia = models.ForeignKey(Terapia, null=True, on_delete=models.CASCADE)
+
+    def __unicode__(self):
+        return self.apellidos + " " + self.nombres
+
+
+
+
+
 
 class Medicamento(models.Model):
     """ Medicamento recetado para convulsiones """
